Trame midi fonctionne

+++ Software/CONTROLEUR_MIDI_CODE/Core/Inc/midi.h
@@ -8,6 +8,6 @@
 #ifndef INC_MIDI_H_
 #define INC_MIDI_H_
 
-void send_new_value(int num_pot, uint16_t new_value);
+void send_new_value(uint8_t num_pot, uint8_t new_value);
 
 #endif /* INC_MIDI_H_ */
